+++ dsl/src/main/groovy/com/readytalk/jenkins/model/PipelineDslException.groovy
@@ -1,0 +1,5 @@
+package com.readytalk.jenkins.model
+
+//TODO: Expand on this - for now, it's just a way for the top-level
+//TODO: methods to inject error context
+class PipelineDslException extends RuntimeException { }